refactor: split indent from render mod
Split the render mod up into logical parts, and move the Indent type
into the indent helper.
